feat: migrate views to use  directory for components
