+++ handlers/register_initial.go
@@ -2,8 +2,8 @@ package handlers
 
 import (
 	"github.com/labstack/echo/v4"
-	"github.com/sonr-io/motr/components/views"
 	"github.com/sonr-io/motr/middleware"
+	"github.com/sonr-io/motr/ui/components/views"
 )
 
 func HandleRegisterInitial(c echo.Context) error {
