+++ ui/components/charts/area_chart.templ
