+++ ui/components/charts/bar_chart.templ
