+++ ui/components/charts/candle_chart.templ
