+++ ui/components/charts/candle_chart_templ.go
