+++ ui/components/charts/line_chart.templ
