+++ ui/components/dashboard/helia_console.templ
