+++ ui/components/dashboard/helia_console_templ.go
