+++ ui/components/inputs/input_avatar.templ
