+++ ui/components/menus/assets_dropdown.templ
