+++ ui/components/menus/assets_dropdown_templ.go
