+++ ui/components/menus/coins_dropdown.templ
