+++ ui/components/views/console.templ
@@ -2,7 +2,7 @@ package views
 
 import (
 	"github.com/sonr-io/motr/ui"
-	"github.com/sonr-io/motr/components/dashboard"
+	"github.com/sonr-io/motr/ui/components/dashboard"
 	"time"
 )
