+++ ui/components/views/register.templ
@@ -2,7 +2,7 @@ package views
 
 import (
 	"github.com/sonr-io/motr/ui"
-	"github.com/sonr-io/motr/components/inputs"
+	"github.com/sonr-io/motr/ui/components/inputs"
 )
 
 templ RegisterView() {
